requirements.txt
certifi==2016.9.26
click==6.6
Flask==0.11.1
future==0.15.2
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-telegram-bot==5.1.1
SQLAlchemy==1.1.2
urllib3==1.18
Werkzeug==0.11.11
